Grant copy implemented for transmit for performance improvement.
Figured that the usage of grant copy, instead of grantmap, provides better
trhoughput. However, ways of using grant map is kept in the code so that we
can switch to grantmap or use both (size bigger than 128 uses map and others
use copy). We are also batching tx operations to a fixed number.